Force all args of std::max to be int32.
This might fix https://trac.osgeo.org/gdal/ticket/6212 for Visual Studio as reported by Joaquim Luis.  It looks like
GDAL is being built as C++11 with std::initializer_list in the error.


git-svn-id: https://svn.osgeo.org/gdal/trunk@31463 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/hdf4/hdf4imagedataset.cpp
@@ -2382,9 +2382,9 @@ int HDF4ImageDataset::ProcessSwathGeolocation( int32 hSW, char **papszDimList )
     }
     else
     {
-        // aim for 10x10 grid or so.
-        iGCPStepX = std::max( 1, ((nXPoints-1) / 11) );
-        iGCPStepY = std::max( 1, ((nYPoints-1) / 11) );
+        // Aim for 10x10 grid or so.
+        iGCPStepX = std::max( static_cast<int32>(1), ((nXPoints-1) / 11) );
+        iGCPStepY = std::max( static_cast<int32>(1), ((nYPoints-1) / 11) );
     }
 
 /* -------------------------------------------------------------------- */
